mgmt-back/requirements.txt
aio-pika
fastapi
ldap3
pymongo
python-dotenv
redis
uvicorn
aio-pika